refactor: 프로젝트 구조 수정
- domain, utils 패키지 추가
